refactor(ui-components): promote rich FilePreviewDialog to @spaarke/ui-components 2.1.0
Moves the SemanticSearchControl PCF's rich `FilePreviewDialog` (2-column metadata
pane, prev/next navigation, 3-dot row-action menu, AI summary support) from
its PCF-local home into `@spaarke/ui-components` as `RichFilePreviewDialog`,
then migrates all 3 consumer locations to use it:

  • PCF SemanticSearchControl (v1.1.69 → v1.1.70): local `FilePreviewDialog.tsx`
    collapses from 998 lines to a 22-line re-export shim that aliases the shared
    `RichFilePreviewDialog` as `FilePreviewDialog`. Zero behavior change.
  • LegalWorkspace: local `FilePreviewDialog.tsx` collapses from 451 lines to a
    197-line thin adapter that wires LW-specific BFF services into the shared
    rich callback API. UX upgrade — LW now gets the 2-column layout + metadata
    pane + prev/next-ready chrome.
  • DocumentRelationshipViewer: local `FilePreviewDialog.tsx` collapses from
    287 lines to a 15-line re-export shim. App.tsx passes `documentId` (newly
    required). UX upgrade matches LegalWorkspace.

The simpler existing shared `FilePreviewDialog` (services-injection API, 880px,
single-column) is marked `@deprecated` but retained — `FindSimilarResultsStep`
still consumes it and migration would cascade across 8+ downstream consumers
(separate refactor).

Net code change: −2,236 lines in consumer-local implementations; +910 lines in
shared lib (the canonical rich dialog). All 4 affected packages build clean.

Architecture impact:
  - Reduces FilePreview parallel implementations from 4 → 2 (canonical rich
    in shared lib + legacy simple in shared lib)
  - Future surfaces (Office Add-ins, future SPAs) inherit the rich UX
    automatically
  - Bug/feature improvements to the preview dialog now ship to all consumers
    via shared-lib version bumps instead of N-way local maintenance

@see projects/spaarke-matter-ui-enhancement-r1/notes/handoffs/handoff-2026-05-29-pre-compact.md
(Section: "v1.1.70+ follow-up queue / Primary / FilePreview promotion")
