+++ src/client/pcf/SemanticSearchControl/SemanticSearchControl/ControlManifest.Input.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <manifest>
-  <control namespace="Sprk" constructor="SemanticSearchControl" version="1.1.69"
+  <control namespace="Sprk" constructor="SemanticSearchControl" version="1.1.70"
            display-name-key="Semantic Search Control"
-           description-key="PCF control for semantic document search with natural language queries (v1.1.69)"
+           description-key="PCF control for semantic document search with natural language queries (v1.1.70)"
            control-type="virtual">
 
     <external-service-usage enabled="true">
