+++ src/client/pcf/SemanticSearchControl/Solution/solution.xml
@@ -11,7 +11,7 @@
     <Descriptions>
       <Description description="PCF control for semantic document search with natural language queries, infinite scroll, and filtering" languagecode="1033" />
     </Descriptions>
-    <Version>1.1.69</Version>
+    <Version>1.1.70</Version>
     <Managed>0</Managed>
     <Publisher>
       <UniqueName>Spaarke</UniqueName>
