+++ src/client/shared/Spaarke.UI.Components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@spaarke/ui-components",
-  "version": "2.0.0",
+  "version": "2.1.0",
   "description": "Spaarke shared UI component library - Fluent UI v9 (selective imports)",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
